% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cM}
\alias{cM}
\title{A random walk Metropolis sampler}
\usage{
cM(sigma = 0.05, x0 = 25, N = 5000L)
}
\arguments{
\item{sigma}{the standard deviation of Normal}

\item{x0}{The initial value}

\item{N}{Generate N random numbers}
}
\value{
a random sample of size \code{N}
}
\description{
A random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
cM(0.05,25,5000)
}
}
