% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{min_gcv}
\alias{min_gcv}
\title{choose tuning parameters in scad by GCV}
\usage{
min_gcv(a, lambda, x, y)
}
\arguments{
\item{a}{vector of tuning parameter a}

\item{lambda}{vector of tuning parameter lambda}

\item{x}{n*p matrix x}

\item{y}{n*1 matrix y}
}
\value{
The a and the lambda that minimize GCV
}
\description{
use GCV to choose tuning parameters when using smoothly clipped absolute deviation (SCAD) for variable selection and estimation.
}
\examples{
\dontrun{
library(MASS)
n=40
d=8
Sig=matrix(0,nrow=d,ncol=d)
for(i in 1:d){
  for (j in 1:d) Sig[i,j]=0.5^abs(i-j)
}
mu = rep(0,d)
x=mvrnorm(n, mu, Sig)
beta=c(3,1.5,0,0,2,0,0,0)
epsilon=rnorm(n,0,3)
y=x\%*\%beta+epsilon
a=seq(3,4,0.1)
lambda=seq(0,1,0.1)
c=min_gcv(a,lambda,x,y)
c
}
}
