\name{StatComp20010-package}
\alias{StatComp20010-package}
\docType{package}
\title{
  Three functions for 'Statistical Computing' course.
}
\description{
  This package includes three main functions. First, a function that uses the spearman rank correlation for ultrahigh dimensional data to reduce the dimension of the predictor variable to a moderate level.Second, a robust sure independence screening for ultrahigh dimensional data and use the hard thresholding rule to reduce the dimension of the predictor variable.Third, a function to use GCV to choose tuning parameters when using smoothly clipped absolute deviation (SCAD) for variable selection and estimation.
}
\author{
Shijie Yang

Maintainer: Shijie Yang <yang666@mail.ustc.edu.cn>
}
\keyword{ package }
